Increase compiler warnings in wireguard_py/PACKAGE
Summary:
LLVM-15 has many useful compiler warnings that are not yet enabled globally. In order to enable them globally we are first enabling them on a per-directory basis.

This diff enables additional warning flags for this directory. If the builds succeed, the diff should be safe to land.

 - If you approve of this diff, please use the "Accept & Ship" button :-)

(1 file modified.)

Reviewed By: dmm-fb

Differential Revision: D52763090

fbshipit-source-id: 79f314005eb57002d47fecc60f12ae58a9fa8ce3